requirements.txt

playwright==1.46.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
Jinja2==3.1.4
pydantic==2.8.2
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
imapclient==2.3.1
beautifulsoup4==4.12.3
html5lib==1.1
